+++ main.cpp
@@ -55,10 +55,14 @@ int main(void)
 
 
     Globals::Game.init();
+    Portal::PortalShader = LoadShader(0,TextFormat("shaders/fragments/portal.h",330));
+
+
 
     Camera3D& camera = Globals::Game.camera;
 
     Globals::Game.loadLevel("levels/layer0.txt");
+    Globals::Game.loadLevel("levels/layer1.txt");
     Globals::Game.setLayer(0);
 
     float accum = 0;