+++ src/factory.cpp
@@ -20,6 +20,14 @@ std::shared_ptr<PhysicsBody> construct(std::string cereal)
         {
             ptr = new PortalSpawner(Factory<PortalSpawner>::Base::deserialize(params));
         }
+        else if (params[0] == Factory<Portal>::ObjectName)
+        {
+            ptr = new Portal(Factory<Portal>::Base::deserialize(params));
+        }
+        else if (params[0] == Factory<Collectible>::ObjectName)
+        {
+            ptr = new Collectible(Factory<Collectible>::Base::deserialize(params));
+        }
         else
         {
             std::cerr << "ERROR construct: unable to construct object: " << params[0] << "\n";