+++ src/objects.cpp
@@ -2,6 +2,11 @@
 
 #include "../headers/objects.h"
 
+void PhysicsBody::setPos(const Vector2& pos)
+{
+    orient.pos = pos;
+}
+
 void PhysicsBody::setDead(bool val)
 {
     dead = val;